split Favorite header component to two different desktop and tablet
