+++ src/components/FavoriteHeaderDesktop/FavoriteHeaderDesktop.jsx
@@ -1,0 +1,9 @@
+import './FavoriteHeaderDesktop.css'
+
+const FavoriteHeaderDesktop = () => {
+    return < div className='favoriteHeaderDesktop' >
+        <p className='favoriteHeaderText'>Favorite</p>
+    </ div>
+}
+
+export default FavoriteHeaderDesktop;
