+++ src/components/FavoriteHeaderDesktop/index.js
@@ -1,0 +1,3 @@
+import FavoriteHeaderDesktop from './FavoriteHeaderDesktop';
+
+export default FavoriteHeaderDesktop;
