+++ src/components/FavoriteHeaderTablet/FavoriteHeaderTablet.css
@@ -1,9 +1,3 @@
-.favoriteHeaderDesktop {
-    padding-right: 10px;
-    margin-top: 10px;
-    margin-bottom: 20px;
-    display: block;
-}
 
 .favoriteHeaderTablet {
     display: none;
@@ -54,11 +48,4 @@
         cursor: pointer;
     }
 
-    .favoriteBlockClosed {
-        display: none;
-    }
-
-    .favoriteAsideOpened {
-        display: block;
-    }
 }
