+++ src/components/FavoriteHeaderTablet/index.js
@@ -1,0 +1,3 @@
+import FavoriteHeaderTablet from './FavoriteHeaderTablet';
+
+export default FavoriteHeaderTablet;
